# Use .NET 5's HEX function