This PR aligns us with .NET and uses [`Convert.FromHexString`](https://learn.microsoft.com/en-us/dotnet/api/system.convert.fromhexstring?view=net-5.0) which is vectorized in .NET, so we should observe significant speedup (see https://github.com/MetacoSA/NBitcoin/pull/1149 for some numbers). 

Current master implementation

```
|  Method |      Mean |    Error |   StdDev |
|-------- |----------:|---------:|---------:|
|   ToHex |  25.19 ns | 0.501 ns | 1.355 ns |
| FromHex | 563.06 ns | 4.742 ns | 3.702 ns |
```

This PR: Only first commit:

```
|  Method |     Mean |    Error |   StdDev |
|-------- |---------:|---------:|---------:|
|   ToHex | 35.58 ns | 0.171 ns | 0.152 ns |
| FromHex | 71.38 ns | 0.650 ns | 0.576 ns |
```

This PR: All commits

```
|  Method |     Mean |    Error |   StdDev |
|-------- |---------:|---------:|---------:|
|   ToHex | 25.80 ns | 0.474 ns | 1.180 ns |
| FromHex | 75.86 ns | 1.552 ns | 1.963 ns |
```


Benchmark is [here](https://github.com/kiminuo/WalletWasabi/tree/feature/2022-11-17-fast-HEX-benchmark), use in `WalletWasabi.Tests` folder:

```s
dotnet run -c Release --framework net6.0 -- --runtimes net6.0
```


With this PR, we can replace https://github.com/zkSNACKs/WalletWasabi/blob/master/WalletWasabi/Backend/Models/FilterModel.cs#L47 to use our new implementation and it can shave a bit of time from startup. Fundamentally, we should tackle it differently, but given current consensus, this is what we can do.