@turbolay 2022-11-17
Reading filters entirely on testnet (`ForEachFilters` function at launched with `fromHeight` at `BlockchainTip`):
Master: `Encoders.Hex.DecodeData` called 1578162 times -> 132 ms
PR: `ByteHelpers.FromHex` called 1578162 -> 81 ms

Same on Mainet:
Master: `Encoders.Hex.DecodeData` called 285851 times -> 656 ms
PR: `ByteHelpers.FromHex` called 285851 times -> 626 ms

I guess that `Encoders.Hex.DecodeData` is already fast